Moved TimeoutHandler, added Copyright text to several classes
