+++ SMTInterpol/src/de/uni_freiburg/informatik/ultimate/smtinterpol/muses/Shrinking.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2020 Leonard Fichtner
+ *
+ * This file is part of SMTInterpol.
+ *
+ * SMTInterpol is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * SMTInterpol is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with SMTInterpol.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package de.uni_freiburg.informatik.ultimate.smtinterpol.muses;
 
 import java.util.ArrayList;
@@ -11,7 +29,7 @@ import de.uni_freiburg.informatik.ultimate.smtinterpol.smtlib2.TerminationReques
 /**
  * A class that provides methods for single MUS extraction.
  *
- * @author LeonardFichtner (leonard.fichtner@web.de)
+ * @author LeonardFichtner
  *
  */
 public class Shrinking {
